chore: fix wasm build
