+++ libs/client-api/src/lib.rs
@@ -2,6 +2,7 @@ mod http;
 mod http_ai;
 mod http_billing;
 
+mod http_blob;
 mod http_collab;
 mod http_history;
 mod http_member;
