+++ libs/client-api/src/native/mod.rs
@@ -1,9 +1,7 @@
-mod http_blob;
 mod http_native;
 mod ping;
 mod retry;
 
-pub use http_blob::*;
 #[allow(unused_imports)]
 pub use http_native::*;
 pub(crate) use ping::*;
